+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+## [2.0.1] - 2026-06-23
+
+### Added
+
+- `utils.ts#RequestError` error class to wrap request `method`, `url` and response `status` code for errors
+  - `response.json()` errors should be simple `TypeError` and are unhandled
+
+### Fixed
+
+- handling of `resourceId` array in `results.ts#postSearch`
+
 ## [2.0.0] - 2026-06-19
 
 ### Changed
@@ -81,7 +92,8 @@ Extracted FCS Aggregator REST API adatper code from [`textplus-fcs-store`](https
 - Modularized typed interfaces with `axios` request methods.
 - Mini test suite. WIP.
 
-[Unreleased]: https://github.com/clarin-eric/fcs-sru-aggregator-api-adapter-typescript/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/clarin-eric/fcs-sru-aggregator-api-adapter-typescript/compare/v2.0.1...HEAD
+[2.0.1]: https://github.com/clarin-eric/fcs-sru-aggregator-api-adapter-typescript/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/clarin-eric/fcs-sru-aggregator-api-adapter-typescript/compare/v1.5.1...v2.0.0
 [1.5.1]: https://github.com/clarin-eric/fcs-sru-aggregator-api-adapter-typescript/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/clarin-eric/fcs-sru-aggregator-api-adapter-typescript/compare/v1.4.0...v1.5.0